+++ containers/tefca-viewer/src/app/customize/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import React, { useState } from "react";
-import CustomizeQuery from "../query/components/CustomizeQuery"; // Adjust the path if necessary
-import { Mode } from "../query/page";
+import CustomizeQuery from "./components/CustomizeQuery"; // Adjust the path if necessary
+import { Mode } from "../constants";
 
 const dummyLabs = [
   {
